 requirements.txt
tensorflow-macos
tensorflow-metal
numpy
matplotlib